Renames everything from mistral to mixtral.
